Delete @LazyInstantiated
This type will come back eventually. Ideally, this functionality would:
- Be receivable by other types down the dependency chain just like @Forwarded and @Instantiated
- Propagate a build warning when you have a dependency that requires instantiating the lazy thing synchronously

The @LazyInstantiated we have today does neither of these things, and
will require breaking changes to make these things happen. So for now,
we remove the type. It's possible for consumers to implement most of
this functionality themselves in the interim utilizing a property
wrapper and @Instantiated. We'll get this right down the line.