Only write the dependency tree output file if it has changed (#167)
@ahmdmhasn had a great idea to not touch the output file if the output
file's content hasn't changed. Looks like modern Xcode seems to support
having a package plugin that does supports this, so let's do it.

I believe earlier SPM implementations required the output file to be
touched as part of the build plugin run, so we _needed_ to write the
file. This does not seem to be the case in Xcode 16+.

---------

Co-authored-by: Ahmed M. Hassan <45182214+ahmdmhasn@users.noreply.github.com>