Instantiable must be conformed to in all isolation contexts (#177)
@mattmassicotte convinced me that it does not make sense for the
`Instantiable` protocol to be conformed to within a single isolation
context. `Instantiable` declares only that a type has a
`ForwardedProperties` type, and this type cannot be scoped to a
particular isolation context, and therefore `Instantiable` _is_ a
`SendableMetatype`.

This PR makes `Instantiable` required to be a `SendableMetatype` in
Swift versions 6.2 and above. I may not be able to merge this PR (due to
code coverage requirements) until a Swift 6.2 linux container is added
to CI. TBD.